Addition of a refined energy strategy
Each agent has an strategy S which include the parameters R (reservoir)
in [0,1] and H (heritage) in [0, energy-zero].  Initially, agents has a
randpm strategy and they transmit it to their descendants. Strategy
works as follows:
* Each agent moves a distance calculated as its energy minus basal
metabolism times (1 - R)  and keeps the fraction of energy described by
R. Then they consume in a circle of radius equal to Vision + (1 - R).
* The agents reproduce if and only if they have an acumulated energy
equals to (H + energy-zero) * 2. In some sense this is like an extra
amount of energy which suposselly improves survivance of descendants.